patch for a 64bit beta version. It compile, and work apart for fork(), i think...

+++ xmview-os/um-viewos/sctab.h
@@ -65,8 +65,8 @@ extern char um_patherror[];
 
 void scdtab_init();
 
-char *um_getpath(int laddr,struct pcb *pc);
-char *um_abspath(int laddr,struct pcb *pc,struct stat64 *pst,int link);
+char *um_getpath(long laddr,struct pcb *pc);
+char *um_abspath(long laddr,struct pcb *pc,struct stat64 *pst,int link);
 
 void um_set_errno(struct pcb *pc,int i);
 char *um_getcwd(struct pcb *pc,char *buf,int size);
